+++ frontend/src/components/list/List.tsx
@@ -1,7 +1,7 @@
 import useList from "hooks/list.hooks/useList";
-import ListReviewButtons from "./ListReviewButtons/ListReviewButtons";
-import ListTerms from "./ListTerms/ListTerms";
-import ListTitleBar from "./ListTitleBar/ListTitleBar";
+import ListReviewButtons from "./sub/ListReviewButtons";
+import ListTerms from "./sub/ListTerms";
+import ListTitleBar from "./sub/ListTitleBar";
 
 function List() {
 	const { list, termsToDisplay, filter, handleListTitleBlur, handleDelete, setFilter } =
