+++ frontend/src/hooks/list.hooks/useList.tsx
@@ -1,4 +1,3 @@
-import ListTerm from "components/list/ListTerms/ListTerm";
 import { numTermsToReviewState } from "components/review/state/reviewSelectors";
 import { List } from "gql/codegen-output";
 import { useQueryListsById } from "gql/hooks/list-query";
@@ -7,6 +6,7 @@ import { useCallback, useEffect, useState } from "react";
 import { useRecoilValue, useResetRecoilState, useSetRecoilState } from "recoil";
 import { listState } from "state/atoms/listAtoms";
 import { TruncatedTerm } from "types/list.types";
+import ListTerm from "../../components/list/sub/ListTerm";
 import { useListFilter } from "./useListFilter";
 import { useListPrepareReview } from "./useListPrepareReview";
 import { useListUpdate } from "./useListUpdate";
